+++ src/FileHelper.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 
 namespace Berlioz\Helpers;
 
+use InvalidArgumentException;
+
 /**
  * Class FileHelper.
  *
@@ -107,5 +109,142 @@ final class FileHelper
                         return (int)$size;
                 }
         }
+    }
+
+    /**
+     * Resolve absolute path.
+     *
+     * @param string $srcPath
+     * @param string $dstPath
+     *
+     * @return string|null
+     */
+    public static function resolveAbsolutePath(string $srcPath, string $dstPath): ?string
+    {
+        $srcPath = self::uniformizePathSeparator($srcPath);
+        $dstPath = self::uniformizePathSeparator($dstPath);
+        $finalPath = $dstPath;
+
+        if (substr($dstPath, 0, 1) !== '/') {
+            // Complete absolute link
+            if (substr($dstPath, 0, 2) === './') {
+                $dstPath = substr($dstPath, 2);
+            }
+
+            // Unification of directories separators
+            $finalPath = self::uniformizePathSeparator(dirname($srcPath));
+            if ($finalPath === '.') {
+                $finalPath = '';
+            }
+
+            // Concatenation
+            $finalPath = sprintf('%s/%s', $finalPath, $dstPath);
+        }
+
+        // Replacement of './'
+        $finalPath = str_replace('/./', '/', $finalPath);
+
+        // Replacement of '../'
+        do {
+            $finalPath = preg_replace('#(/|^)([^\\\/?%*:|"<>.]+)/../#', '/', $finalPath, -1, $nbReplacements);
+        } while ($nbReplacements > 0);
+
+        if (false === strpos($finalPath, './')) {
+            return '/' . ltrim($finalPath, '/');
+        }
+
+        return null;
+    }
+
+    /**
+     * Resolve relative path.
+     *
+     * @param string $srcPath
+     * @param string $dstPath
+     *
+     * @return string
+     */
+    public static function resolveRelativePath(string $srcPath, string $dstPath): string
+    {
+        $srcPath = ltrim(self::resolveAbsolutePath('/', $srcPath), '/');
+        $dstPath = ltrim(self::resolveAbsolutePath($srcPath, $dstPath), '/');
+
+        if (substr($srcPath, 0, 2) === '..') {
+            throw new InvalidArgumentException('Source path must be a relative path');
+        }
+        if (substr($srcPath, 0, 2) === './') {
+            $srcPath = substr($srcPath, 2);
+        }
+
+        $srcPath = explode('/', $srcPath);
+        $dstPath = explode('/', $dstPath);
+
+        // Already relative?
+        if (in_array(reset($dstPath), ['.', '..'])) {
+            return implode('/', $dstPath);
+        }
+
+        // Get filename of destination path
+        $dstFilename = self::extractFilename($dstPath);
+        self::extractFilename($srcPath);
+
+        $srcDepth = count($srcPath);
+        $dstDepth = count($dstPath);
+        $differentDepthPath = 0;
+
+        for ($i = 0; $i < $srcDepth; $i++) {
+            if (!isset($dstPath[$i]) || $srcPath[$i] !== $dstPath[$i] || $differentDepthPath > 0) {
+                $differentDepthPath++;
+            }
+        }
+
+        $relativePath = '';
+        if ($differentDepthPath > 0) {
+            $relativePath .= str_repeat('../', $differentDepthPath);
+            $relativePath .= implode('/', array_slice($dstPath, min($dstDepth, $differentDepthPath) - 1));
+        }
+        if ($differentDepthPath === 0) {
+            $relativePath .= './';
+            $relativePath .= implode('/', array_slice($dstPath, $srcDepth, $dstDepth));
+        }
+
+        // Add file to relative path
+        if (null !== $dstFilename) {
+            $relativePath .= '/' . $dstFilename;
+        }
+
+        return preg_replace('#/{2,}#', '/', $relativePath);
+    }
+
+    /**
+     * Uniformize path separator.
+     *
+     * @param string $path
+     *
+     * @return string
+     */
+    private static function uniformizePathSeparator(string $path): string
+    {
+        $path = str_replace(['\\', '/'], '/', $path);
+
+        return preg_replace('#/{2,}#', '', $path);
+    }
+
+    /**
+     * Extract filename.
+     *
+     * @param array $path
+     *
+     * @return string|null
+     */
+    private static function extractFilename(array &$path): ?string
+    {
+        $filename = end($path) ?: null;
+
+        if (null !== $filename) {
+            unset($path[count($path) - 1]);
+        }
+
+        return $filename;
     }
 }