+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.5.0] - 2021-12-22
+
+### Added
+
+- New method `FileHelper::resolveAbsolutePath()` to resolve absolute path from another
+- New method `FileHelper::resolveRelativePath()` to resolve relative path from another
+
 ## [1.4.0] - 2021-12-07
 
 ### Changed